requirements.txt
Django==4.2.18
requests==2.32.3
djangorestframework==3.15.2
drf-yasg==1.21.8

